requirements.txt
python-telegram-bot[persistence]
python-telegram-bot[job-queue]
requests
python-dotenv
timezonefinder
numpy
Flask